ADD_상가 세부 정보 통신, 예약 취소 *** 서버 재배포 되면 수정 필요

+++ Travely_User/app/src/main/java/com/hyeran/android/travely_user/model/ReviewResponseData.kt
@@ -2,7 +2,7 @@ package com.hyeran.android.travely_user.model
 
 data class ReviewResponseData (
         var content : String,
-        var createdAt : Int,
+        var createdAt : Double,
         var like : Int,
         var reviewIdx : Int,
         var storeIdx : Int,
